Breaking Change - package reorganization
Moved `gotrader` package under `gotrader/gotrader`. The motivation is that theroot folder is getting messy, and there is confusion between gotrader packages and external tools, like gotaset
