+++ gotrader/alpaca/alpaca_test.go
@@ -2,7 +2,7 @@ package alpacabroker
 
 import (
 	_ "github.com/joho/godotenv/autoload"
-	"github.com/totomz/gotrader"
+	"github.com/totomz/gotrader/gotrader"
 	"log"
 	"os"
 	"testing"
