+++ gotrader/cerbero_test.go
