+++ gotrader/datafeed.go
