+++ gotrader/datafeed_test.go
