+++ gotrader/signals.go
