+++ gotrader/signals_test.go
