+++ gotrader/strategy.go
