+++ gotrader/strategy_test.go
